Sequential training cycles
You can now run multiple training cyles in a row. You can set a fixed number or allow it to run forever.

After each cycle, the performance of the new modelset vs the starting point is evaluated. If it is not better, the starting point is used again. If the new modelset is better, it is used as the starting point for the next cycle.

Graceful early stopping is handled by deleting a file "delete_to_stop_training_cycles.txt" [jank!].

Prettier printing of game playing progress.