Changed how models are handled to better work with convolutional models.
Training data is stored as some shape with as many dimensions as makes sense (connect 4 has 3 with a 4th for rows). When training models with this data, it is first reshaped to match the needs of the model. This target shape is stored in the AITrainer struct.

Resolved an issue where loaded models could not be used. The issue appears to be that they included an anonymous function which could not be loaded. This was resolved by naming that function.

Resolved an issue with the calculation of model loss during training. The loss function used the weights of the model before training to calculate y_hat, rather than the updated weights after each epic. This is fixed.