Changed how model input sizes are tracked.
There is now a struct, Modelset, which contains the models for every game phase and the input sizes of those models. If an input size is not specified for a model, it is assumed that input data's shape does not need to be changed (this could be errouneous). The advantage of this change is to reduce the number of arguments passed to methods which use the models.

Minor change: when evaulating the quality of a newly trained modelset vs an old one, ties are counted as half for each player.